refactor:#48 useEffect를 useQuery로 분리
